backend/requirements.txt
# ─── Web framework ────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# ─── Async task queue ─────────────────────────────────────────────────────────
celery[redis]==5.3.6
redis==5.0.4

# ─── HTTP client ──────────────────────────────────────────────────────────────
httpx==0.27.0
tenacity==8.3.0

# ─── Data science ─────────────────────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.2
scipy==1.13.0
biopython==1.83
scikit-learn==1.5.0

# ─── Storage ──────────────────────────────────────────────────────────────────
minio==7.2.7

# ─── Config / Validation ──────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ─── WebSockets ───────────────────────────────────────────────────────────────
websockets==12.0

# ─── Authentication ───────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
bcrypt==4.0.1
cryptography>=41.0.0

# ─── AI / LLM ─────────────────────────────────────────────────────────────────
anthropic>=0.25.0

# ─── PDF generation ───────────────────────────────────────────────────────────
reportlab>=4.0.0

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
